% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_res_time.R
\name{gg_res_time}
\alias{gg_res_time}
\title{residual vs time plot}
\usage{
gg_res_time(df, .time, .res, .threshold = 2.5, .ct = TRUE)
}
\arguments{
\item{df}{data}

\item{.time}{name of time column}

\item{.res}{name of residual column}

\item{.threshold}{absolute threshold for coloring outliers, defaults to 2.5}

\item{.ct}{show central tendency curve, default TRUE}
}
\description{
residual vs time plot
}
\details{
this function is primarily used for CWRES vs TAD
}
\examples{
\dontrun{
   df \%>\% gg_res_time(TAD, CWRES)
}
}
