% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_element.R
\name{last_element}
\alias{last_element}
\title{get the last element of a vector}
\usage{
last_element(.x)
}
\arguments{
\item{.x}{vector}
}
\description{
get the last element of a vector
}
\examples{
last_element(c(1, 7, 6, 4))
}
